Implement method 'get_segment' in order to get each single segment within a domain.
The domain classes and their 'create_segments' method have been modified to store the created segments within the class in an STD vector. This is pretty convenient, because it allows the 'get_segment' method to be implemented using a simple random acces to the position where the proper segment can be found in the vector.

Overloaded constructors for the segment classes have been defined, in order to be able to construct segment objects from a ViennaGrid segment, which is needed by the modifications introduced in the domain classes and the 'create_segments' method.