Implement method 'get_vertex' in order to get each single vertex stored within a domain.
Overloaded constructors for the point classes have been defined, in order to be able to construct point objects from a ViennaGrid point, which is needed by the modifications introduced in the domain classes in order to support the 'get_vertex' method.

The equal operator for points had to be implemented in order to be able to run assertions that compare if two vertices are equal.