Implement method 'create_cell' for cell creation within a segment.
Segments have been modified to also store a pointer to the domain to which the segment belongs. This is necessary for 'create_cell' to be able to get the domain from which the vertices of the cell are to be extracted.

Domains have been modified to allow it to return a reference to the ViennaGrid domain, so that 'create_cell' can know from what domain the vertices of the cell are to be extracted. The 'create_segments' method of domains has been modified to pass a reference to the domain when initializing segments, for the reason stated above.