Improve performance of getting properties 'segments' and 'vertices'.
Previously, the list of segments and the list of vertices where being stored in an STL vector within the class. Every time any of the properties was being requested, the whole list had to be iterated over many times, once for copying from the STL vector to a Python list, and then again when returning the Python list (by copy).

The changes introduced in this commit consist in storing the list directly as a Python list, instead of using an STL vector. This way, both 'create_segments' and 'add_vertex' "append" to a Python list instead of "pushing back" to an STL vector. 'create_segments' is hence O(n) as it was previously. However, 'get_segments' and 'get_vertices' now do not need to copy the STL vector to a Python list, thus reducing the temporal complexity from O(2n) to O(1n).

Note that in the theory of big-O notation, the coefficient does not make a difference. However, we use it here to indicate that the execution time with O(2n) will be in practice twice as long as with O(1n), because the same O(n) loop is being executed twice (thus coefficient 3) instead of just once (coefficient 1).