Add constructors to cartesian points to construct points from ViennaGrid points.
The old constructor that was getting a pointer to the ViennaGrid point has been modified to accept a pointer as parameter, instead of a reference, in order to avoid ambiguity.

The new constructor accepts an object by copy and reserves memory in the heap segment to store a copy of the passed point.