Fix error that caused calls on methods of the class 'Cell' to raise an exception.
Since the method 'create_cell' of the wrapped segments didn't return the actual cell (it was a 'void' method), Python interpreted the return type as 'None' and, thus, the method 'Segment.create_cell' of the high-level module initialized a new 'Cell' from a 'None' object. Hence, the methods of the class 'Cell' failed when trying to access any attribute or method of the wrapped cell.

To solve this, the method 'create_cell' of the wrapped segments have been refactored to return the new cell they create instead of 'void'.