Implement VTK writer for linear domains with segmentations.
Previously, only the VTK writer for linear domains without segmentations was implemented, i.e., the writer only writes the information contained in the domain object, but no segmentation object is passed to the writer and thus no segmentation information is written to the file.

This commit implements the writer with segmentations, i.e., a segmentation object can be passed to the writer, so that the writer writes both domain and segmentation information to the file.

The way this has been implemented is the following: Having the old writer without segmentations, an optional parameter has been added to the writer method in order to be able to pass to the method a pointer to a segmentation object. If the parameter is present, the writer with segmentations is used. Otherwise, the writer without segmentations is used.

In order to implement this using Boost.Python, we had to use the overloading capabilities of Boost.Python described in [1] and combine this documentation with the code exmample of [2], since the information in [1] is not complete.

[1]: http://www.boost.org/doc/libs/1_54_0/libs/python/doc/tutorial/doc/html/python/functions.html#python.default_arguments
[2]: http://wiki.python.org/moin/boost.python/FunctionOverloading