Fix algorithm 'cell_refine'.
Since the third parameter of the function was a pointer to a C++ function, trying to apply the 'cell_refine' algorithm passing a Python function as the third parameter caused an 'ArgumentError' to be raised by Boost.Python.

To be able to use this algorithm, the type of the third parameter had to be changed to a Python object, since Python functions are also objects and must be handled in Boost.Python as such. This way, the Python object that represents a Python function can be called with any other Python objects as parameters, even in C++. That's how we call the 'predicate' function onto the cell object to be processed.